Add AddPlace component and change NewPlacePage wrapper page.
